+++ test/Test/Toml/Codec/Combinator/Common.hs
@@ -1,5 +1,6 @@
 module Test.Toml.Codec.Combinator.Common
     ( codecRoundtrip
+    , exactCodecRoundtrip
 
       -- * Double helpers
     , Batman (..)
@@ -8,12 +9,14 @@ module Test.Toml.Codec.Combinator.Common
     , batmanFloatCodec
     ) where
 
+import Data.Text (Text)
 import Hedgehog (Gen, forAll, tripping)
 import Test.Hspec (Arg, Expectation, SpecWith, it)
 import Test.Hspec.Hedgehog (hedgehog)
 
 import Toml.Codec.BiMap (TomlBiMap)
-import Toml.Codec.Code (decode, encode)
+import Toml.Codec.Code (decode, decodeExact, encode)
+import Toml.Codec.Error (TomlDecodeError)
 import Toml.Codec.Types (TomlCodec)
 import Toml.Type.AnyValue (AnyValue)
 import Toml.Type.Key (Key)
@@ -22,6 +25,22 @@ import Toml.Type.Key (Key)
 import qualified Toml.Codec as Toml
 
 
+codecRoundtripWith
+    :: forall a
+    .  (Eq a, Show a)
+    => (TomlCodec a -> Text -> Either [TomlDecodeError] a)
+    -> String
+    -> (Key -> TomlCodec a)
+    -> Gen a
+    -> SpecWith (Arg Expectation)
+codecRoundtripWith dcode typeName mkCodec genA = it label $ hedgehog $ do
+    a <- forAll genA
+    let codec = mkCodec "a"
+    tripping a (encode codec) (dcode codec)
+  where
+    label :: String
+    label = typeName ++ ": decode . encode ≡ id"
+
 codecRoundtrip
     :: forall a
     .  (Eq a, Show a)
@@ -29,13 +48,16 @@ codecRoundtrip
     -> (Key -> TomlCodec a)
     -> Gen a
     -> SpecWith (Arg Expectation)
-codecRoundtrip typeName mkCodec genA = it label $ hedgehog $ do
-    a <- forAll genA
-    let codec = mkCodec "a"
-    tripping a (encode codec) (decode codec)
-  where
-    label :: String
-    label = typeName ++ ": decode . encode ≡ id"
+codecRoundtrip = codecRoundtripWith decode
+
+exactCodecRoundtrip
+    :: forall a
+    .  (Eq a, Show a)
+    => String
+    -> (Key -> TomlCodec a)
+    -> Gen a
+    -> SpecWith (Arg Expectation)
+exactCodecRoundtrip str = codecRoundtripWith decodeExact ("Exact " <> str)
 
 -- | Wrapper over 'Double' and 'Float' to be equal on @NaN@ values.
 newtype Batman a = Batman